Move migrations to WriteRepo
Also, keep the old Repo’s migrations directory to prevent this error when booting the app:

```
** (Mix) Could not find migrations directory "priv/repo/migrations"
for repo Gatekeeper.Repo.
This may be because you are in a new project and the
migration directory has not been created yet. Creating an
empty directory at the path above will fix this error.
If you expected existing migrations to be found, please
make sure your repository has been properly configured
and the configured path exists.
```