Remove direct Member <-> Door Group association
This feature was the ability to directly grant Members access to DoorGroups. The idea was to enable certain people within a company to have more access to doors than others in the same company. However, this caused more confusion that it was worth (since there was a temptation to check those boxes when setting up members), and never actually was needed in production over the last 2 years. It also made the unlock logic significantly more complex.